A text classifier that predicts which one of two classes of texts to which a string belongs.
In order to run, two textfiles named CodeReference and TextReference are needed.
When a reference code and a reference text are provided, the classifier can be trained to recognize wether a
string is computer code  or text. Half of each provided file is used as training data and half as evaluation data.